requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
doctran==0.0.14
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.50
language_data==1.2.0
lxml==4.9.4
marisa-trie==1.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==0.27.10
orjson==3.10.1
packaging==23.2
phonenumbers==8.13.35
pillow==10.3.0
pluggy==1.5.0
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==1.10.15
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
pyproject-toml==0.0.10
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
requests-file==2.0.0
response==0.5.0
rpds-py==0.18.0
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.5.2
tldextract==5.1.2
toml==0.10.2
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
wheel==0.43.0
yarl==1.9.4
